requirements/base.txt
django
djangorestframework
django-filter
django-cors-headers
django-suit
django-extensions
dj-database-url
dj-static
django-rest-swagger
